requirements.txt
# oTree-may-not-overwrite-this-file
otree>=6.0.0
psycopg2-binary>=2.8.4
boto3>=1.42.28
httpx>=0.27.0
python-dotenv>=1.0.0
pandas>=2.0.0
sentry-sdk>=2.0.0